+++ client/src/components/Navbar/navbar.styles.ts
@@ -16,11 +16,9 @@ export const Content = styled.div`
     padding: 10px 0;
     margin: 0 auto; //center the content div itself
     color: black;
-
     a {
         text-decoration: none;
     }
-
     @media screen and (max-width: 720px){
       padding: 10px 0;   //decrease size of header on smaller screens
     }
@@ -28,14 +26,12 @@ export const Content = styled.div`
 
 export const LogoImg = styled.img`
     width: 100px;
-
     .text {
         font-family: 'Gotham', sans-serif;
        // color: black;
         padding-left: 20px;
         font-size: 500;
     }
-
     //when the screen size is less than 720px
     @media screen and (max-width: 720px){
       width: 120px;   
@@ -45,30 +41,25 @@ export const LogoImg = styled.img`
 export const NavLinks = styled.div`
     display: flex;
     align-items: center;
-
     /* menu links section */
     .nav-links {
         padding: 0 15px;
     }
-
     .nav-link {
         font-size: var(--fontSmedium);
         font-weight: 400;
     }
-
     .nav-link:hover {
         color: var(--slate);
         border-bottom: 2px solid var(--darkRed);
         padding-bottom: 16px;
     }
-
     .active {
         color: var(--slate);
         border-bottom: 3px solid var(--darkRed);
         padding-bottom: 16px;
         font-weight: 410;
     }
-
     /* login and search icons section*/
     .nav-icons{
         border-left: 2px solid var(--fog);
@@ -77,11 +68,8 @@ export const NavLinks = styled.div`
       
     img {
         width: 35px;
-
         @media screen and (max-width: 720px) {
             width: 40px;
         }
     }
-
-
 `;
