+++ client/src/pages/Album/album.styles.ts
