+++ client/src/pages/Photos/photos.styles.ts
@@ -115,3 +115,5 @@ export const ThumbsContainer = styled.aside`
     }
 `;
 
+
+
