+++ client/src/pages/Register/register.styles.ts
