+++ client/src/pages/Upload/upload.styles.ts
